Made a new Class "userHistory" to manage user saves
