+++ WpfApplication1/FTPClient.csproj
@@ -74,6 +74,7 @@
     </ApplicationDefinition>
     <Compile Include="removeDir.cs" />
     <Compile Include="removeFile.cs" />
+    <Compile Include="userHistory.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -118,6 +119,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="userHistory.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
